pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "genre-classifier"
version = "0.1.0"
description = "GTZAN genre classifier with SVM and CNN baselines"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Sam Witt"}]
dependencies = [
    "numpy>=1.24",
    "librosa>=0.10",
    "soundfile>=0.12",
    "scikit-learn>=1.3",
    "tensorflow>=2.13",
    "tensorflow-datasets>=4.9",
    "matplotlib>=3.7",
    "pydantic>=1.10",
    "PyYAML>=6.0",
]

[project.scripts]
genre-classify = "genre_classifier.cli:main"
